more testing and document editing done

+++ src/main/java/com/project/service/ItemsService.java
@@ -44,18 +44,25 @@ public ItemsDTO create(ItemsTable entity) {
 
 
 public boolean delete(long id) {
-	  ItemsTable ent = this.repo.findById(id).orElseThrow(IDNotFoundExceptions::new);
+	if(!this.repo.existsById(id)) {
+		throw new IDNotFoundExceptions();
+	}
+	else {
+	  //ItemsTable ent = this.repo.findById(id).orElseThrow(IDNotFoundExceptions::new);
       this.repo.deleteById(id);
       boolean isExist = this.repo.existsById(id);
       return isExist; 
-     
+	}
  }
 public boolean deleteUniqueID(long id) {
-	ItemsTable ent = this.repo.findAllItemsBySQL(id);
+	if(this.repo.findItemByUniqueIDSQL(id) == null) {
+		throw new IDNotFoundExceptions();
+	}
+	else {
     this.repo.deleteItemBySQL(id);
     boolean isExist = this.repo.existsById(id);
     return isExist; 
-   
+	}
 }
 
 public ItemsDTO update(long id, ItemsTable entity) {
@@ -74,7 +81,7 @@ public String createUni(ItemsTable entity) {
 	long range = 9999999L;
 	long number = (long)(rand.nextDouble()*range);
     this.repo.saveItemBySQL(entity.getItemName(),entity.getPrice(),entity.getStock(), number);
-    return "Coin created with Id: " + number; 
+    return "Item created"; 
 }
 
 }
