+++ src/main/java/com/project/service/userService.java
@@ -42,10 +42,15 @@ public AcDTO create(UserTable entity) {
 
 
 public boolean delete(long id) {
-	  UserTable ent = this.repo.findById(id).orElseThrow(IDNotFoundExceptions::new);
+	if(!this.repo.existsById(id)) {
+		throw new IDNotFoundExceptions();
+	}
+	else {
+
       this.repo.deleteById(id);
       boolean isExist = this.repo.existsById(id);
       return isExist; 
+	}
      
  }
